Mobile: expanded rows show the full point detail; drawer fetch guard keyed per request
The mobile table's expand-on-tap previously only unfolded the text. The
detail body of the desktop drawer (images with click-to-original lightbox,
prominent text block, per-datatype field list) is now extracted into a
shared PointDetailContent component and rendered inside the expanded row,
with a chevron affordance and tap-header-to-collapse. Expanding a row also
grows the bottom sheet (up to min(520px, 60vh)) so the detail is visible
without dragging first. Rows use the already-fetched dataTableRows objects,
so no extra request is made. The lightbox intercepts Escape in the capture
phase so an enclosing drawer doesn't also close.

Review follow-up: the drawer's stale-response guard now keys on
dataset.id + sae.id + index instead of index alone, so a slow response
from a previously viewed dataset can no longer overwrite the drawer when
the row number happens to match.
